refactor(Control): change data type for command arguments to update a product.

+++ AquaEngine.API/Control/Domain/Model/Commands/UpdateProductOwnerCommand.cs
@@ -1,3 +1,3 @@
 ﻿namespace AquaEngine.API.Control.Domain.Model.Commands;
 
-public record UpdateProductOwnerCommand(long UserId, long ProductId);
+public record UpdateProductOwnerCommand(int UserId, int ProductId);
